requirements.txt
Flask==2.3.2
gunicorn==20.1.0 
psycopg2-binary==2.9.9
sqlalchemy==2.0.29
python-dotenv==1.0.1
Flask==2.3.2